Sync before importing utxos if chain tip is greater than current
There is a condition that utxos are being added at the time that bitcoind
is adding a new block to the chain, and we've yet to add that block to
the current wallets, we will first need to update the status of the wallets
before being able to add the utxos sets. Unlike the getaddressdeltas method
from bitcoind that can query between any two sets of block heights, the
getaddressutxos query will always return results as of the latest block
height. Thus we need to make sure that the tip is what is expected.

Another alternative is to include prevTxId and prevOut in the
getaddressdeltas bitcoind RPC command, so that it's able to reconstruct
the utxo set for the wallet based on those deltas, and to know which
outputs have been spent.